server/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
psycopg2-binary
torchcodec
supabase
pyjwt
bcrypt
boto3
email-validator
python-multipart
requests
langchain-openai
langchain-core
langchain-openai
langchain-core
langchain-google-genai
